+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 * Add support for returning multiple messages by a list of message IDs
+* Improved `Delta` support: added `Delta` model and two new methods; `since` and `longpoll`
 
 ### 6.1.0 / 2022-01-28
 * Add support for `Event` to ICS